Rename package `module` files to `index.js`
Makes it easier to import, you don't have to specify the file name, just
the package name.
